backend/requirements.txt
Django
djangorestframework
django-cors-headers
scikit-learn
joblib
numpy
phonenumbers
reportlab
xhtml2pdf